We were only partially using warm key We were still passing KEY_NAME (cold key) in the joins, and since we created the COLD key on those nodes and are sharing the keychain, we were getting away with it. This fixes that by changing ONLY the KEY_NAME in the `api` container for joins by adding -WARM. then it also fixes up sending the grant command, which was all wrong since it was expecting it to be a transaction (since it was in `tx` CLI namespace), but it's more of a command.
