+++ local-test-net/docker-compose.join.yml
@@ -14,3 +14,4 @@ services:
       - ACCOUNT_PUBKEY=${ACCOUNT_PUBKEY}
       - KEYRING_BACKEND=${KEYRING_BACKEND:-test}
       - KEYRING_PASSWORD=${KEYRING_PASSWORD}
+      - KEY_NAME=${KEY_NAME}-WARM
