+++ testermint/src/main/kotlin/DockerGroup.kt
@@ -38,7 +38,7 @@ data class GenesisUrls(val keyName: String) {
 
 data class DockerGroup(
     val dockerClient: DockerClient,
-    val keyName: String,
+    val pairName: String,
     val publicPort: Int,
     val mlPort: Int,
     val adminPort: Int,
@@ -51,16 +51,18 @@ data class DockerGroup(
     val workingDirectory: String,
     val genesisGroup: GenesisUrls? = null,
     val genesisOverridesFile: String,
-    val publicUrl: String = "http://$keyName-api:9000",
-    val pocCallbackUrl: String = "http://$keyName-api:9100",
+    val publicUrl: String = "http://$pairName-api:9000",
+    val pocCallbackUrl: String = "http://$pairName-api:9100",
     val config: ApplicationConfig,
     val useSnapshots: Boolean,
 ) {
+    val warmKeyName = "$pairName-WARM"
+    val coldKeyName = pairName
     val composeFiles = when (isGenesis) {
         true -> GENESIS_COMPOSE_FILES
         false -> NODE_COMPOSE_FILES
     }.let { baseFiles: List<String> ->
-        val additionalFiles = config.additionalDockerFilesByKeyName[keyName] ?: emptyList()
+        val additionalFiles = config.additionalDockerFilesByKeyName[pairName] ?: emptyList()
         baseFiles + additionalFiles.map { "$LOCAL_TEST_NET_DIR/$it" }
     }.onEach { file: String ->
         if (!Path.of(workingDirectory, file).exists()) {
@@ -75,18 +77,18 @@ data class DockerGroup(
             .also { it.environment().putAll(envMap) }
     }
 
-    val warmKeyPassword = this.keyName.padEnd(10, '0')
+    val warmKeyPassword = this.pairName.padEnd(10, '0')
 
     // return the pubkey for the cold key
     fun createColdKey(): String {
         val command = listOf(
             "docker", "compose",
-            "-p", keyName
+            "-p", pairName
         ) + composeFiles.flatMap { listOf("-f", it) } + listOf(
             "--project-directory", workingDirectory,
             "run", "--rm", "--no-deps", "api",
             "sh", "-c",
-            """printf '%s\n%s\n' "${warmKeyPassword}" "${warmKeyPassword}" | inferenced keys add $keyName --keyring-backend file"""
+            """printf '%s\n%s\n' "${warmKeyPassword}" "${warmKeyPassword}" | inferenced keys add $coldKeyName --keyring-backend file"""
         )
 
         val process = ProcessBuilder(command)
@@ -110,12 +112,12 @@ data class DockerGroup(
     fun createWarmKey(): String {
         val command = listOf(
             "docker", "compose",
-            "-p", keyName
+            "-p", pairName
         ) + composeFiles.flatMap { listOf("-f", it) } + listOf(
             "--project-directory", workingDirectory,
             "run", "--rm", "--no-deps", "api",
             "sh", "-c",
-            """printf '%s\n%s\n' "${warmKeyPassword}" "${warmKeyPassword}" | inferenced keys add $keyName-WARM --keyring-backend file"""
+            """printf '%s\n%s\n' "${warmKeyPassword}" "${warmKeyPassword}" | inferenced keys add $warmKeyName --keyring-backend file"""
         )
 
         val process = ProcessBuilder(command)
@@ -141,12 +143,12 @@ data class DockerGroup(
             createWarmKey()
             accountPubkey
         } else ""
-        val composeArgs = mutableListOf("compose", "-p", keyName)
+        val composeArgs = mutableListOf("compose", "-p", pairName)
         composeFiles.forEach { file ->
             composeArgs.addAll(listOf("-f", file))
         }
         composeArgs.addAll(listOf("--project-directory", workingDirectory))
-        composeArgs.addAll(listOf( "up", "-d"))
+        composeArgs.addAll(listOf("up", "-d"))
         val baseArgs = composeArgs.toImmutableList()
         if (!isGenesis) {
             // This will allow us to get our consensus key and add the participant BEFORE we launch the API
@@ -161,10 +163,24 @@ data class DockerGroup(
             Thread.sleep(Duration.ofSeconds(10))
 
             val containers = getRawContainers(config)
-            val node = containers.getCli(this.keyName) ?: error("Could not find node container for keyName=${this.keyName}")
-            Thread.sleep(Duration.ofSeconds(5))
-            val validatorKey = node.getValidatorInfo().key
-            node.registerNewParticipant(publicUrl, accountPubKey, validatorKey, this.genesisGroup?.apiUrl ?: "http://genesis-api:9000")
+            val node =
+                containers.getCli(this.pairName) ?: error("Could not find node container for keyName=${this.pairName}")
+            val validatorKey = (1..3).fold<Int, String?>(null) { acc, _ ->
+                acc ?: try {
+                    node.getValidatorInfo().key
+                } catch (e: com.google.gson.JsonSyntaxException) {
+                    Logger.warn("Validator key not yet available, waiting 5 seconds and trying again", "")
+                    Thread.sleep(Duration.ofSeconds(5))
+                    null
+                }
+            } ?: throw IllegalStateException("Failed to get validator info after 3 attempts")
+            node.registerNewParticipant(
+                publicUrl,
+                accountPubKey,
+                validatorKey,
+                this.genesisGroup?.apiUrl ?: "http://genesis-api:9000"
+            )
+            node.grantMlOpsPermissionsToWarmAccount()
             val startRemainingArgs = baseArgs + listOf("api", "mock-server", "proxy")
             this.coldAccountPubkey = node.getColdPubKey()
             dockerProcess(*startRemainingArgs.toTypedArray()).start().waitFor()
@@ -185,8 +201,8 @@ data class DockerGroup(
     }
 
     fun tearDownExisting() {
-        Logger.info("Tearing down existing docker group with keyName={}", keyName)
-        val composeArgs = mutableListOf("compose", "-p", keyName)
+        Logger.info("Tearing down existing docker group with keyName={}", pairName)
+        val composeArgs = mutableListOf("compose", "-p", pairName)
         composeFiles.forEach { file ->
             composeArgs.addAll(listOf("-f", file))
         }
@@ -198,15 +214,18 @@ data class DockerGroup(
 
     private fun getCommonEnvMap(useSnapshots: Boolean): Map<String, String> {
         return buildMap {
+            put("KEY_NAME", coldKeyName)
             coldAccountPubkey?.let {
                 put("ACCOUNT_PUBKEY", it)
                 put("KEYRING_BACKEND", "file")
                 put("KEYRING_PASSWORD", warmKeyPassword)
                 put("CREATE_KEY", "false")
+                // KEY_NAME in our docker/compose files is used as pair-name a LOT. We will need to unwind this
+                // For now, docker-compose.join.yml adds "-WARM" to the env variable only.
+//                put("KEY_NAME", warmKeyName)
             }
-            put("KEY_NAME", keyName)
             put("KEYRING_PASSWORD", warmKeyPassword)
-            put("NODE_HOST", "$keyName-node")
+            put("NODE_HOST", "$pairName-node")
             put("DAPI_API__POC_CALLBACK_URL", pocCallbackUrl)
             put("DAPI_API__PUBLIC_URL", publicUrl)
             put("DAPI_API__PUBLIC_SERVER_PORT", "9000")
@@ -273,9 +292,9 @@ data class DockerGroup(
             }
         }
 
-        val inferenceDir = baseDir.resolve("prod-local/$keyName")
-        val mappingsDir = baseDir.resolve("prod-local/mock-server/$keyName/mappings")
-        val filesDir = baseDir.resolve("prod-local/mock-server/$keyName/__files")
+        val inferenceDir = baseDir.resolve("prod-local/$pairName")
+        val mappingsDir = baseDir.resolve("prod-local/mock-server/$pairName/mappings")
+        val filesDir = baseDir.resolve("prod-local/mock-server/$pairName/__files")
         val mappingsSourceDir = baseDir.resolve("testermint/src/main/resources/mappings")
         val publicHtmlDir = baseDir.resolve("public-html")
 
@@ -286,7 +305,7 @@ data class DockerGroup(
 
         val templatePath = "testermint/src/main/resources/alternative-mappings/validate_poc_batch.template.json"
         val templateContent = baseDir.resolve(templatePath).toFile().readText()
-        val content = templateContent.replace("{{KEY_NAME}}", keyName)
+        val content = templateContent.replace("{{KEY_NAME}}", pairName)
         val mappingFile = mappingsDir.resolve("validate_poc_batch.json")
         Files.writeString(mappingFile, content)
 
@@ -295,7 +314,7 @@ data class DockerGroup(
         }
         val jsonOverrides = config.genesisSpec?.toJson(cosmosJson)?.let { "{ \"app_state\": $it }" } ?: "{}"
         Files.writeString(inferenceDir.resolve("genesis_overrides.json"), jsonOverrides, StandardOpenOption.CREATE)
-        Logger.info("Setup files for keyName={}", keyName)
+        Logger.info("Setup files for keyName={}", pairName)
     }
 
     init {
@@ -337,7 +356,7 @@ fun createDockerGroup(
     }
     return DockerGroup(
         dockerClient = DockerClientBuilder.getInstance().build(),
-        keyName = keyName,
+        pairName = keyName,
         publicPort = 9000 + iteration,
         mlPort = 9001 + iteration,
         adminPort = 9002 + iteration,
@@ -374,7 +393,7 @@ fun initializeCluster(joinCount: Int = 0, config: ApplicationConfig, currentClus
                 createDockerGroup(
                     index + 1,
                     actualIndex,
-                    GenesisUrls(genesisGroup.keyName.trimStart('/')),
+                    GenesisUrls(genesisGroup.pairName.trimStart('/')),
                     config,
                     false
                 )
@@ -382,7 +401,7 @@ fun initializeCluster(joinCount: Int = 0, config: ApplicationConfig, currentClus
         }
         val joinGroups = (1..joinCount).mapIndexed { index, _ ->
             val actualIndex = (index + 1) * 10
-            createDockerGroup(index + 1, actualIndex, GenesisUrls(genesisGroup.keyName.trimStart('/')), config, false)
+            createDockerGroup(index + 1, actualIndex, GenesisUrls(genesisGroup.pairName.trimStart('/')), config, false)
         }
         val allGroups = listOf(genesisGroup) + joinGroups
         Logger.info("Initializing cluster with {} nodes", allGroups.size)
